+++ test/memory/contextKlassTests.cpp
@@ -1,7 +1,6 @@
-# include "incls/_precompiled.incl"
-# include "incls/_BlockKlass.cpp.incl"
-#include "handle.hpp"
-#include "test.h"
+#include "easyunit/test.h"
+#include "memory/markSweep.hpp"
+#include "oops/klassOop.hpp"
 
 using namespace easyunit;
 
@@ -25,7 +24,7 @@ TEARDOWN(ContextKlassTests){
 
 TESTF(ContextKlassTests, allocateShouldFailWhenAllowedAndNoSpace) {
   eden_top = eden_end;
-  ASSERT_EQUALS((int)NULL, (int)(theClass->klass_part()->allocateObject(false)));
+  ASSERT_EQUALS((intptr_t)NULL, (intptr_t)(theClass->klass_part()->allocateObject(false)));
 }
 
 TESTF(ContextKlassTests, allocateShouldAllocateTenuredWhenRequired) {
